feat: run sync hook
