+++ src/hooks/sync-loop-hook.ts
@@ -8,19 +8,10 @@ export default class SyncLoopHook<
     for (let i = 0; i < this.callbacks.length; i++) {
       const name = this.options[i].name;
       const callback = this.callbacks[i];
-      try {
-        let result: Return | void | undefined;
-        do {
-          result = callback(...args);
-        } while (result !== undefined);
-      } catch (err) {
-        const e = err as Error;
-        throw this.createError(e.message, {
-          type: 'call',
-          receiver: name,
-          stack: e.stack,
-        });
-      }
+      let result: Return | void | undefined;
+      do {
+        result = this.runCallback(name, callback, args);
+      } while (result !== undefined);
     }
   }
 }
