+++ src/hooks/sync-waterfall-hook.ts
@@ -8,17 +8,8 @@ export default class SyncWaterfallHook<
     for (let i = 0; i < this.callbacks.length; i++) {
       const name = this.options[i].name;
       const callback = this.callbacks[i];
-      try {
-        const waterfall = callback(...newArgs);
-        if (waterfall !== undefined) newArgs[0] = waterfall;
-      } catch (err) {
-        const e = err as Error;
-        throw this.createError(e.message, {
-          type: 'call',
-          receiver: name,
-          stack: e.stack,
-        });
-      }
+      const result = this.runCallback(name, callback, newArgs);
+      if (result !== undefined) newArgs[0] = result;
     }
     return newArgs[0];
   }
